+++ includes/modules/class-sales-module.php
@@ -37,6 +37,9 @@ class Sales_Module extends Module_Base {
 	protected string $id   = 'sales';
 	protected string $name = 'Sales';
 
+	protected string $exclusive_group = 'commerce';
+	protected int $exclusive_priority = 10;
+
 	/**
 	 * Sync direction: Sales module only pulls from Odoo.
 	 *